AlbionBotGather/requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cx-Freeze==6.14.9
cx-Logging==3.1.0
cycler==0.11.0
filelock==3.12.0
fonttools==4.39.3
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
keyboard==0.13.5
kiwisolver==1.4.4
lief==0.13.0
MarkupSafe==2.1.2
matplotlib==3.7.1
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.1
Nuitka==1.5.8
numpy==1.24.3
opencv-python==4.7.0.72
ordered-set==4.1.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
psutil==5.9.5
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.5
pytz==2023.3
PyYAML==6.0
requests==2.29.0
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.21.1
six==1.16.0
sympy==1.11.1
thop==0.1.1.post2209072238
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.90
urllib3==1.26.15
zipp==3.15.0
zstandard==0.21.0
